Well the tests fail but at least it doesn't immediately error
