+++ addon/templates/components/basic-power-select.hbs
@@ -1,0 +1,85 @@
+{{#basic-dropdown
+  horizontalPosition=(readonly horizontalPosition)
+  calculatePosition=calculatePosition
+  destination=(readonly destination)
+  initiallyOpened=(readonly initiallyOpened)
+  matchTriggerWidth=(readonly matchTriggerWidth)
+  onClose=(action "onClose")
+  onOpen=(action "onOpen")
+  registerAPI=(action "registerAPI")
+  renderInPlace=(readonly renderInPlace)
+  verticalPosition=(readonly verticalPosition)
+  disabled=(readonly disabled)
+as |dropdown|}}
+  {{yield
+    (hash
+      dropdown=(hash
+        trigger=(component dropdown.trigger
+          tagName=(readonly _triggerTagName)
+          ariaDescribedBy=(readonly ariaDescribedBy)
+          ariaInvalid=(readonly ariaInvalid)
+          ariaLabel=(readonly ariaLabel)
+          ariaLabelledBy=(readonly ariaLabelledBy)
+          ariaRequired=(readonly required)
+          class=(readonly concatenatedTriggerClasses)
+          id=(readonly triggerId)
+          onKeyDown=(action "onTriggerKeydown")
+          onFocus=(action "onTriggerFocus")
+          onBlur=(action "onTriggerBlur")
+          tabindex=(readonly tabindex)
+        )
+        content=(component dropdown.content
+          _contentTagName=_contentTagName
+          class=(readonly concatenatedDropdownClasses)
+        )
+      )
+      before-options=(component beforeOptionsComponent
+        extra=(readonly extra)
+        listboxId=(readonly optionsId)
+        onInput=(action "onInput")
+        onKeydown=(action "onKeydown")
+        searchEnabled=(readonly searchEnabled)
+        onFocus=(action "onFocus")
+        onBlur=(action "onBlur")
+        searchPlaceholder=(readonly searchPlaceholder)
+        select=(readonly publicAPI)
+      )
+      options=(component optionsComponent
+        class="ember-power-select-options"
+        extra=(readonly extra)
+        groupIndex=""
+        loadingMessage=(readonly loadingMessage)
+        id=(readonly optionsId)
+        options=(readonly publicAPI.results)
+        optionsComponent=(readonly optionsComponent)
+        groupComponent=(readonly groupComponent)
+        select=(readonly publicAPI)
+      )
+      placeholder=(component 'power-select/placeholder')
+      power-select-group=(component 'power-select/power-select-group')
+      search-message=(component searchMessageComponent
+        searchMessage=(readonly searchMessage)
+        select=(readonly publicAPI)
+      )
+      trigger=(component triggerComponent
+        allowClear=(readonly allowClear)
+        buildSelection=(readonly buildSelection)
+        extra=(readonly extra)
+        listboxId=(readonly optionsId)
+        onFocus=(action "onFocus")
+        onBlur=(action "onBlur")
+        onInput=(action "onInput")
+        placeholder=(readonly placeholder)
+        placeholderComponent=(readonly placeholderComponent)
+        onKeydown=(action "onKeydown")
+        searchEnabled=(readonly searchEnabled)
+        searchField=(readonly searchField)
+        select=(readonly publicAPI)
+        selectedItemComponent=(readonly selectedItemComponent)
+      )
+      after-options=(component afterOptionsComponent
+        select=(readonly publicAPI) extra=(readonly extra)
+      )
+    )
+  }}
+{{/basic-dropdown}}
