+++ app/components/basic-power-select.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-power-select/components/basic-power-select';
