+++ tests/integration/components/basic-power-select-test.js
@@ -1,0 +1,6 @@
+import { moduleForComponent } from 'ember-qunit';
+// import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('basic-power-select', 'Integration | Component | basic power select', {
+  integration: true
+});
